% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalEvid_FullBayesian.R
\name{MarginalEvid_FullBayesian}
\alias{MarginalEvid_FullBayesian}
\title{MarginalEvid}
\usage{
MarginalEvid_FullBayesian(
  Data_Y,
  Data_time,
  nu0 = 1,
  Lambda0 = 1,
  Sigma0 = 1,
  Beta0 = NULL
)
}
\arguments{
\item{Data_Y}{This is observed multi-responses (nsamples x nsites)}

\item{Data_time}{This is observed times (nsamples vector)}

\item{nu0}{degrees of freedom hyperparameter of invWishart prior distribution (should be 0=non-informative or 1=informative)}

\item{Lambda0}{Scale matrix hyperparameter of invWishart prior distribution of Covariance (can be numeric or matrix)}

\item{Sigma0}{Covariance hyperparameter of Normal prior distribution of beta (can be numeric, 2 or 3 long vector)}

\item{Beta0}{Mean hyperparameter of Normal prior distribution of beta (can be 2 long vector or matrix)}
}
\value{
The numerical marginal evidence
}
\description{
Calculating the marginal evidence for the data
}
\details{
This function provides the marginal evidence value based on all the data

 Model: A full Bayesian model (takes into account the uncertainty of the parameters)
 p(Y|time) = int_theta p(Y|time,theta)p(theta) d_theta
 Where
 p(Y|time,theta)=MVN(b0 + b1*time, SIGMA), SIGMA= COVARIANCE MATRIX
 with latent priors 'normal-invWishart'
 p(b0,b1|SIGMA)=MVN(0,0, diag(sig0 SIGMA,sig1 SIGMA))
 p(SIGMA) = invWishart(Lambda0,nu0), where diag(Lambda0) = lam0

Marginal evidence p(Y|eta) = int_theta p(Y|theta)p(theta|eta)d theta
}
\examples{
\dontrun{ 
n = 100
dat = genData(n,seed=1)
Data_Y = dat$Data_Y
Data_time = dat$Data_time
modelEvid = MarginalEvid_FullBayesian(Data_Y, Data_time)
}
}
\author{
Oyvind Bleka
}
