% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredictions.R
\name{plotPredictions}
\alias{plotPredictions}
\title{plotPredictions}
\usage{
plotPredictions(
  predObj,
  TrueTimes = NULL,
  text_main = "",
  text_xlab = "time",
  text_ylab = "Probability density function",
  pdfFileName = NULL
)
}
\arguments{
\item{predObj}{An object returned from either posteriorTime_NaiveMLE or posteriorTime_FullBayesian}

\item{TrueTimes}{Vector of times (true). Must be same length as the number of samples}

\item{text_main}{A text used to show in plot}

\item{text_xlab}{A text used to show in plot}

\item{text_ylab}{A text used to show in plot}

\item{pdfFileName}{File name of pdf to print (NULL means no pdf is created)}
}
\description{
Plot Time distributions for predicted individuals
}
\details{
The function creates plots for showing TSD predictions
}
